The list of available solvers is now generated by cmake and stored in a header.
As a result, the real list of headers available is
known to the final user. If he tries to use an
unknown - or unavailable - solver, the system will
crash during the compilation of the project rather
that at the execution.

This commit will not compile, because of the cmake list system.