make tests compatible with qibuild
It boils down to replacing

  ADD_TEST(${name} ${name})

with

  ADD_TEST(${name} ${fullpath/to/test/exec/file})

So that qibuild can find the test it should run. However, this needs special
casing for visual studio, where the place exec files lies is not easy to
know by default.

Note that this fix is not specific to qibuild, I do not think ctest would find
the tests without it too.

Note also that jrl-cmake does something similar, but by reading the code, I
do not think it works, they seem to use a property as a variable.