harmonize reinclusion guard macros
They are now following this pattern:

  MPC_WALKGEN_<DIRNAME>_<FILENAMEWITHEXTENSION>

With -./ replaced with _.

The <DIRNAME> part is optional, it is used to make a difference between
files existing in several directories (common/zebulon/humanoid).

Note that before this commit, some re-inclusion guards looked like this:

  #ifndef _MPC_WALKGEN_QP_SOLVER_TYPE_H_
   ...

I think this is wrong, as names starting with _ followed by an upper-case
letter are reserved for the C++ implementation (in any namespace).

I also added #pragma once.