Revert "FindLSSOL: fix LSSOL segfault due to wrong link order"
This reverts commit de9d174b2f9750c6d300e5d922a70de91656aa59.

This commit restores the previous link order. We had a problem
with it on ia32, but it was most probably the result of a flacky
LSSOL build.

Now, with a freshly rebuilt LSSOL (as a static library), and the
restored link order, it works on both 32 and 64 bits versions of linux.

So with this commit

   ia32, static libs: ok
   x64, static libs: ok

Note that the LSSOL build uses both f2c and blas from the ubuntu
packages, not the ones included in LSSOL distribution.